chore: adding info message about port forwarding
Depending on setup you might need to fight with iptables
and your cloud provider to allow network traffic through to
the application. This is out of scope for this install
script since the process is different for each provider.

This little reminder tells you it might be required, but
that you will have to figure this out on your own.